fix(ui) standardize keyboard handling across app
feat: set android.softwareKeyboardLayoutMode to 'resize'
create WWKeyboardAvoidingView component
refactor: replace KeyboardAvoidingView in EngineerConsoleScreen
refactor: update WWScreenView to use native android resize
chore: remove legacy CustomKeyboardAvoidingView
