+++ src/database/schema.ts
@@ -12,7 +12,7 @@
 import { appSchema, tableSchema } from '@nozbe/watermelondb'
 
 export default appSchema({
-    version: 158,
+    version: 159,
     tables: [
         tableSchema({
             name: 'activity_sensitivity',
